cursor-ide-autopilot/requirements.txt
flask
openai>=1.0.0
pytesseract
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
pytest
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
pyautogui==0.9.54
flask-limiter>=3.5.0
flask-jwt-extended>=4.5.3
psutil>=5.9.0
memory-profiler>=0.61.0
Pillow>=10.2.0
watchdog